refactor gfssf
