+++ ballet/validation/feature_acceptance/validator.py
@@ -25,18 +25,19 @@ class GFSSFAcceptanceEvaluator(FeatureAcceptanceEvaluator):
         self.lmbda_1 = lmbda_1
         self.lmbda_2 = lmbda_2
 
-    def judge(self, feature):
-        feature_df = feature.as_dataframe_mapper() \
-            .fit_transform(self.X_df, self.y)
-        n_samples, n_feature_cols = feature_df.shape
+    def judge(self, candidate_feature):
         feature_dfs_by_src = {}
-        for accepted_feature in self.features:
-            accepted_df = accepted_feature.as_dataframe_mapper().fit_transform(
+        for feature in [candidate_feature] + self.features:
+            feature_df = feature.as_dataframe_mapper().fit_transform(
                 self.X_df, self.y)
-            feature_dfs_by_src[accepted_feature.source] = accepted_df
+            feature_dfs_by_src[feature.source] = feature_df
+
+        candidate_source = candidate_feature.source
+        candidate_df = feature_dfs_by_src[candidate_source]
+        n_samples, n_candidate_cols = feature_df.shape
 
         lmbda_1, lmbda_2 = _compute_lmbdas(
-            self.lmbda_1, self.lmbda_2, feature_df, feature_dfs_by_src)
+            self.lmbda_1, self.lmbda_2, feature_dfs_by_src)
 
         logger.info(
             'Judging Feature using GFSSF: lambda_1={l1}, lambda_2={l2}'.format(
@@ -46,8 +47,10 @@ class GFSSFAcceptanceEvaluator(FeatureAcceptanceEvaluator):
             logger.debug(
                 'Testing with omitted feature: {}'.format(
                     omit or 'None'))
-            z = _concat_datasets(feature_dfs_by_src, n_samples, omit)
-            cmi = estimate_conditional_information(feature_df, self.y, z)
+            z = _concat_datasets(
+                feature_dfs_by_src, n_samples, [
+                    candidate_source, omit])
+            cmi = estimate_conditional_information(candidate_df, self.y, z)
             logger.debug(
                 'Conditional Mutual Information Score: {}'.format(cmi))
             cmi_omit = 0
@@ -60,7 +63,7 @@ class GFSSFAcceptanceEvaluator(FeatureAcceptanceEvaluator):
                 logger.debug('Omitted CMI Score: {}'.format(cmi_omit))
             statistic = cmi - cmi_omit
             threshold = _compute_threshold(
-                lmbda_1, lmbda_2, n_feature_cols, n_omit_cols)
+                lmbda_1, lmbda_2, n_candidate_cols, n_omit_cols)
             logger.debug('Calculated Threshold: {}'.format(threshold))
             if statistic >= threshold:
                 logger.debug(
